Fix build and docs.
