+++ project-management-openproject/pom.xml
@@ -25,7 +25,7 @@
   <parent>
     <groupId>com.xwiki.projectmanagement</groupId>
     <artifactId>project-management</artifactId>
-    <version>1.2.1-SNAPSHOT</version>
+    <version>1.2.21-SNAPSHOT</version>
   </parent>
   <packaging>pom</packaging>
   <name>Project Management - Open Project - POM</name>
